--- tests/Browser/Select/Components/Common/StyledMultipleComponent.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace Tests\Browser\Select\Components\Common;
-
-use Livewire\Component;
-
-class StyledMultipleComponent extends Component
-{
-    public ?array $array = null;
-
-    public function render(): string
-    {
-        return <<<'HTML'
-        <div>
-            {{ implode(',', $array ?? []) }}
-
-            <x-select.styled wire:model="array"
-                             label="Select"
-                             hint="Select"
-                             :options="[
-                                ['label' => 'foo', 'value' => 'foo'],
-                                ['label' => 'bar', 'value' => 'bar'],
-                                ['label' => 'baz', 'value' => 'baz'],
-                             ]"
-                             select="label:label|value:value"
-                             multiple
-            />
-            
-            <x-button dusk="sync" wire:click="sync">Sync</x-button>
-        </div>
-        HTML;
-    }
-
-    public function sync(): void
-    {
-        // ...
-    }
-}
